+++ include/defs.h
@@ -6,7 +6,7 @@
   typedef uint8_t                 u8;
   typedef uint16_t                u16;
   typedef uint32_t                u32;
-  typedef unsigned long long      u64;
+  typedef uint64_t                u64;
   typedef __uint128_t             u128;
 
   #define ALIGN(x)                __attribute__ ((aligned (x)))
